Lazyloading

+++ kproj/chrisOSKernel/src/kInit.c
@@ -123,16 +123,16 @@ __asm__("cli\n");
     
     //isrSavedStack = (uint32_t*)allocPagesAndMapI(kKernelTask->tss->CR3,0x1000); //1k saved stack area
     isrSavedStack = exceptionSavedStack;
-    pagingMapPage(KERNEL_CR3, isrSavedStack,isrSavedStack,0x7);
-    pagingMapPage(KERNEL_CR3, (uint32_t)isrSavedStack | 0xC0000000,(uintptr_t)isrSavedStack,0x7);
+    pagingMapPage(KERNEL_CR3, isrSavedStack,isrSavedStack,(uint16_t)0x7);
+    pagingMapPage(KERNEL_CR3, (uint32_t)isrSavedStack | 0xC0000000,(uintptr_t)isrSavedStack,(uint16_t)0x7);
 
     printd(DEBUG_TASK,"Set up isrSavedStack at 0x%08x (and in the kernel)\n",isrSavedStack);
     
-    pagingMapPage(kKernelTask->tss->CR3,(uintptr_t)kKernelTask->tss,(uintptr_t)kKernelTask->tss,0x7);
-    pagingMapPage(KERNEL_CR3,(uintptr_t)kKernelTask->tss,(uintptr_t)kKernelTask->tss,0x7);
+    pagingMapPage(kKernelTask->tss->CR3,(uintptr_t)kKernelTask->tss,(uintptr_t)kKernelTask->tss,(uint16_t)0x7);
+    pagingMapPage(KERNEL_CR3,(uintptr_t)kKernelTask->tss,(uintptr_t)kKernelTask->tss,(uint16_t)0x7);
     printd(DEBUG_TASK,"Mapped tss of the first task run (0x%08x) into task and kernel\n", kKernelTask->tss);
     
-    pagingMapPage(KERNEL_CR3,PROCESS_STRUCT_VADDR, (uint32_t)kKernelProcess & 0xFFFFF000,0x7); //CLR 1/10/2019: Changed perms from 0x7 to 0x5
+    pagingMapPage(KERNEL_CR3,PROCESS_STRUCT_VADDR, (uint32_t)kKernelProcess & 0xFFFFF000,(uint16_t)0x7); //CLR 1/10/2019: Changed perms from 0x7 to 0x5
     printd(DEBUG_PROCESS,"pagingMapPage(0x%08x,0x%08x,0x%08x,0x%02X)\n",KERNEL_CR3, PROCESS_STRUCT_VADDR, (uint32_t)kKernelProcess & 0xFFFFF000, 0x7);
     gdtEntryOS(kKernelTask->taskNum,(uint32_t)kKernelTask->tss,sizeof(tss_t), ACS_TSS | ACS_DPL_0,GDT_GRANULAR | GDT_32BIT,true);
     __asm__("ltr ax\n"::"a" (kKernelTask->taskNum << 3));
